build: add separate hcache dir
