+++ hcache/backend.h
