+++ hcache/hcache.h
