+++ po/POTFILES.in
@@ -40,13 +40,13 @@ gnupgparse.c
 group.c
 handler.c
 hash.c
-hcache.c
-hcache_bdb.c
-hcache_gdbm.c
-hcache_kc.c
-hcache_lmdb.c
-hcache_qdbm.c
-hcache_tc.c
+hcache/hcache.c
+hcache/bdb.c
+hcache/gdbm.c
+hcache/kc.c
+hcache/lmdb.c
+hcache/qdbm.c
+hcache/tc.c
 hdrline.c
 headers.c
 help.c
